+++ skbio/core/alignment/__init__.py
@@ -20,15 +20,28 @@ Data Structures
    SequenceCollection
    Alignment
 
-Alignment Algorithms
---------------------
+Optimized (i.e., production-ready) Alignment Algorithms
+-------------------------------------------------------
 
 .. autosummary::
    :toctree: generated/
 
    StripedSmithWaterman
    AlignmentStructure
-   align_striped_smith_waterman
+   local_pairwise_align_ssw
+
+Slow (i.e., educational-purposes only) Alignment Algorithms
+-----------------------------------------------------------
+
+.. autosummary::
+   :toctree: generated/
+
+   pairwise.global_pairwise_align_nucleotide
+   pairwise.global_pairwise_align_protein
+   pairwise.global_pairwise_align
+   pairwise.local_pairwise_align_nucleotide
+   pairwise.local_pairwise_align_protein
+   pairwise.local_pairwise_align
 
 Data Structure Examples
 -----------------------
@@ -56,28 +69,22 @@ Data Structure Examples
 >>> s1
 <SequenceCollection: n=2; mean +/- std length=26.50 +/- 1.50>
 
-Alignment Algorithm Examples
-----------------------------
-Using the convenient ``align_striped_smith_waterman`` function:
 
->>> from skbio.core.alignment import align_striped_smith_waterman
->>> alignment = align_striped_smith_waterman(
+Optimized Alignment Algorithm Examples
+--------------------------------------
+Using the convenient ``local_pairwise_align_ssw`` function:
+
+>>> from skbio.core.alignment import local_pairwise_align_ssw
+>>> alignment = local_pairwise_align_ssw(
 ...                 "ACTAAGGCTCTCTACCCCTCTCAGAGA",
 ...                 "ACTAAGGCTCCTAACCCCCTTTTCTCAGA"
 ...             )
 >>> print alignment
-{
-    'optimal_alignment_score': 27,
-    'suboptimal_alignment_score': 21,
-    'query_begin': 0,
-    'query_end': 24,
-    'target_begin': 0,
-    'target_end_optimal': 28,
-    'target_end_suboptimal': 12,
-    'cigar': '10M1I2M1D5M4D7M',
-    'query_sequence': 'ACTAAGGCTCTCTACCCCTCTCAGAGA',
-    'target_sequence': 'ACTAAGGCTCCTAACCCCCTTTTCTCAGA'
-}
+>query
+ACTAAGGCTCTC-TACCC----CTCTCAGA
+>target
+ACTAAGGCTC-CTAACCCCCTTTTCTCAGA
+<BLANKLINE>
 
 Using the ``StripedSmithWaterman`` object:
 
@@ -85,18 +92,10 @@ Using the ``StripedSmithWaterman`` object:
 >>> query = StripedSmithWaterman("ACTAAGGCTCTCTACCCCTCTCAGAGA")
 >>> alignment = query("AAAAAACTCTCTAAACTCACTAAGGCTCTCTACCCCTCTTCAGAGAAGTCGA")
 >>> print alignment
-{
-    'optimal_alignment_score': 49,
-    'suboptimal_alignment_score': 24,
-    'query_begin': 0,
-    'query_end': 26,
-    'target_begin': 18,
-    'target_end_optimal': 45,
-    'target_end_suboptimal': 29,
-    'cigar': '20M1D7M',
-    'query_sequence': 'ACTAAGGCTCTCTACCCCTCTCAGAGA',
-    'target_sequence': 'AAAAAACTCTCTAAACTCACTAAGGCTCTCTACCCCTCTTCAGAGAAGTCGA'
-}
+ACTAAGGCTC...
+ACTAAGGCTC...
+Score: 49
+Length: 28
 
 Using the ``StripedSmithWaterman`` object for multiple targets in an efficient
 way and finding the aligned sequence representations:
@@ -116,22 +115,69 @@ way and finding the aligned sequence representations:
 ...     alignments.append(alignment)
 ...
 >>> print alignments[0]
-{
-    'optimal_alignment_score': 38,
-    'suboptimal_alignment_score': 14,
-    'query_begin': 0,
-    'query_end': 26,
-    'target_begin': 4,
-    'target_end_optimal': 32,
-    'target_end_suboptimal': 15,
-    'cigar': '3M1I6M3D17M',
-    'query_sequence': 'ACTAAGGCTCTCTACCCCTCTCAGAGA',
-    'target_sequence': 'GCTAACTAGGCTCCCTTCTACCCCTCTCAGAGA'
-}
->>> print alignments[0].get_aligned_query_sequence()
+ACTAAGGCT-...
+ACT-AGGCTC...
+Score: 38
+Length: 30
+>>> print alignments[0].aligned_query_sequence
 ACTAAGGCT---CTCTACCCCTCTCAGAGA
->>> print alignments[0].get_aligned_target_sequence()
+>>> print alignments[0].aligned_target_sequence
 ACT-AGGCTCCCTTCTACCCCTCTCAGAGA
+
+Slow Alignment Algorithm Examples
+---------------------------------
+scikit-bio also provides pure-Python implementations of Smith-Waterman and
+Needleman-Wunsch alignment. These are much slower than the methods described
+above, but serve as useful educational examples as they're simpler to
+experiment with. Functions are provided for local and global alignment of
+protein and nucleotide sequences. The ``global*`` and ``local*`` functions
+differ in the underlying algorithm that is applied (``global*`` uses Needleman-
+Wunsch while ``local*`` uses Smith-Waterman), and ``*protein`` and
+``*nucleotide`` differ in their default scoring of matches, mismatches, and
+gaps.
+
+Here we locally align a pair of protein sequences using gap open penalty
+of 11 and a gap extend penalty of 1 (in other words, it is much more
+costly to open a new gap than extend an existing one).
+
+>>> from skbio.core.alignment.pairwise import local_pairwise_align_protein
+>>> s1 = "HEAGAWGHEE"
+>>> s2 = "PAWHEAE"
+>>> r = local_pairwise_align_protein(s1, s2, 11, 1)
+
+This returns an ``skbio.Alignment`` object. We can look at the aligned
+sequences:
+
+>>> print(str(r[0]))
+AWGHE
+>>> print(str(r[1]))
+AW-HE
+
+We can identify the start and end positions of each aligned sequence
+as follows:
+
+>>> r.start_end_positions()
+[(4, 8), (1, 4)]
+
+And we can view the score of the alignment using the ``score`` method:
+
+>>> r.score()
+25.0
+
+Similarly, we can perform global alignment of nucleotide sequences, and print
+the resulting alignment as fasta records:
+
+>>> from skbio.core.alignment.pairwise import global_pairwise_align_nucleotide
+>>> s1 = "GCGTGCCTAAGGTATGCAAG"
+>>> s2 = "ACGTGCCTAGGTACGCAAG"
+>>> r = global_pairwise_align_nucleotide(s1, s2)
+>>> print(r.to_fasta())
+>0
+GCGTGCCTAAGGTATGCAAG
+>1
+ACGTGCCTA-GGTACGCAAG
+<BLANKLINE>
+
 
 """
 
@@ -145,10 +191,10 @@ ACT-AGGCTCCCTTCTACCCCTCTCAGAGA
 
 from .alignment import Alignment, SequenceCollection
 from .ssw.ssw_wrapper import (
-    StripedSmithWaterman, AlignmentStructure, align_striped_smith_waterman)
+    StripedSmithWaterman, local_pairwise_align_ssw, AlignmentStructure)
 
 __all__ = ['Alignment', 'SequenceCollection', 'StripedSmithWaterman',
-           'AlignmentStructure', 'align_striped_smith_waterman']
+           'AlignmentStructure', 'local_pairwise_align_ssw']
 
 from numpy.testing import Tester
 test = Tester().test