+++ CHANGELOG.md
@@ -2,10 +2,34 @@
 
 ## Version 0.1.3-dev (changes since 0.1.3 release go here)
 
+This is a pre-alpha release. At this stage, major backwards-incompatible API changes can and will happen.
+
+### Features
+
+* Added Python implementations of Smith-Waterman and Needleman-Wunsch alignment as ``skbio.core.alignment.pairwise.local_pairwise_align`` and ``skbio.core.alignment.pairwise.global_pairwise_align``. These are much slower than native C implementations (e.g., ``skbio.core.alignment.local_pairwise_align_ssw``) and as a result raise an ``EfficencyWarning`` when called, but are included as they serve as useful educational examples as they’re simple to experiment with.
+* Added ``skbio.core.diversity.beta.pw_distances`` and ``skbio.core.diversity.beta.pw_distances_from_table``. These provide convenient access to the ``scipy.spatial.distance.pdist`` *beta diversity* metrics from within scikit-bio. The ``skbio.core.diversity.beta.pw_distances_from_table`` function will only be available temporarily, until the ``biom.table.Table`` object is merged into scikit-bio (see [#489](https://github.com/biocore/scikit-bio/issues/489)), at which point ``skbio.core.diversity.beta.pw_distances`` will be updated to use that. 
+
+### Backward-incompatible changes
+
+* Function ``skbio.core.alignment.align_striped_smith_waterman`` renamed to ``local_pairwise_align_ssw`` and now returns an ``Alignment`` object instead of an ``AlignmentStructure``
+* The following keyword-arguments for ``StripedSmithWaterman`` and ``local_pairwise_align_ssw`` have been renamed:
+    * ``gap_open`` -> ``gap_open_penalty``
+    * ``gap_extend`` -> ``gap_extend_penalty``
+    * ``match`` -> ``match_score``
+    * ``mismatch`` -> ``mismatch_score``
+* Removed ``skbio.util.sort`` module in favor of [natsort](https://pypi.python.org/pypi/natsort) package.
+
 ### Miscellaneous
 
 * Added powertrip.py script to perform basic sanity-checking of the repo based on recurring issues that weren't being caught until release time; added to Travis build.
 * Added RELEASE.md with release instructions.
+* Added intersphinx mappings to docs so that "See Also" references to numpy, scipy, matplotlib, and pandas are hyperlinks.
+* The following classes are no longer ``namedtuple`` subclasses (see [#359](https://github.com/biocore/scikit-bio/issues/359) for the rationale):
+    * ``skbio.math.stats.ordination.OrdinationResults``
+    * ``skbio.math.gradient.GroupResults``
+    * ``skbio.math.gradient.CategoryResults``
+    * ``skbio.math.gradient.GradientANOVAResults``
+* Added coding guidelines draft.
 
 ## Version 0.1.3 (2014-06-12)